move inplace tests to outside tests folder
