+++ common/meta/raft-store/src/state_machine/database_lookup.rs
@@ -83,27 +83,3 @@ impl fmt::Display for DatabaseLookupValue {
         write!(f, "{}", self.0)
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use common_meta_sled_store::SledOrderedSerde;
-
-    use crate::state_machine::DatabaseLookupKey;
-
-    #[test]
-    fn test_db_lookup_key_serde() {
-        {
-            let k = DatabaseLookupKey::new("tenant1".to_string(), "db1".to_string());
-            let ser_k = k.ser().unwrap();
-            let de_k = DatabaseLookupKey::de(ser_k).unwrap();
-            assert_eq!(k, de_k);
-        }
-
-        {
-            let k = DatabaseLookupKey::new("tenant1".to_string(), "".to_string());
-            let ser_k = k.ser().unwrap();
-            let de_k = DatabaseLookupKey::de(ser_k).unwrap();
-            assert_eq!(k, de_k);
-        }
-    }
-}
